refactor(single): extract history, actions_block, modules_block from mod.rs
Moves build_history_trend_block → history.rs,
build_actions_block → actions_block.rs, and
build_modules_block_from_normalized + build_overall_score_explanation → modules_block.rs.
mod.rs is now a thin entry point (~616 lines) with only build_view_model and tests.
No logic changes.
